requirements.txt
attrs==25.3.0
black==22.1.0
blue==0.9.1
certifi==2025.1.31
click==8.1.8
et_xmlfile==2.0.0
flake8==4.0.1
greenlet==3.2.0
h11==0.14.0
idna==3.10
isort==6.0.1
mccabe==0.6.1
mypy-extensions==1.0.0
numpy==2.2.4
openpyxl==3.1.5
outcome==1.3.0.post0
pandas==2.2.3
pathspec==0.12.1
platformdirs==4.3.7
psycopg2-binary==2.9.10
pycodestyle==2.8.0
pyflakes==2.4.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
selenium==4.31.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.40
tomli==2.2.1
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
websocket-client==1.8.0
wsproto==1.2.0
xlrd==2.0.1
